% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{CPP_EBMA}
\alias{CPP_EBMA}
\title{EBMA Calculation Using C++}
\usage{
CPP_EBMA(x, y, W_Hats, sd, tol)
}
\arguments{
\item{x}{Numeric matrix of  predictions from each model.}

\item{y}{Numeric vector of actual (observed) outcomes.}

\item{W_Hats}{Numeric vector of the weight assigned to each model.}

\item{sd}{Standard deviation.}

\item{tol}{Tolerance threshold after which iteration ends.}
}
\description{
EBMA Calculation Using C++
}
