% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{W_H}
\alias{W_H}
\title{W-Hat Re-weighting Computation.}
\usage{
W_H(z)
}
\arguments{
\item{z}{Numeric matrix of z-hat values, derived in Z hat function}
}
\description{
W-Hat Re-weighting Computation.
}
