% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{Z_H}
\alias{Z_H}
\title{Z-Hat Computation.}
\usage{
Z_H(x, y, w, sd)
}
\arguments{
\item{x}{Numeric matrix of model predictions.}

\item{y}{Vector of actual (observed) outcomes for each observation.}

\item{w}{Vector of weights with each element corresponding to a model.}

\item{sd}{Standard deviation.}
}
\description{
Z-Hat Computation.
}
