% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{W_H}
\alias{W_H}
\title{W-Hat Re-weighting Computation.}
\usage{
W_H(Z_H)
}
\arguments{
\item{Z_H}{Numeric matrix of z-hat values, derived in Z_H function}
}
\description{
W-Hat Re-weighting Computation.
}
